ClassMethods#[] accepts an IPAddr::Base
Previously, IPAddr::V4.[] would accept an IPAddr input, but would not
preserve the prefix length. With this change, the prefix length is
preserved.

ClassMethods#[] was a bit unwieldy already, so rather than make it
worse, several stand-alone conversion functions have been split out to
support this change:

* IPAddr::Base.from_integer
* IPAddr::Base.from_ipaddr
* IPAddr::Base.from_string
* IPAddr::Base.object_to_prefix_length
* IPAddr::Base.integer_to_prefix_length
* IPAddr::Base.ipaddr_to_prefix_length
* IPAddr::Base.string_to_prefix_length

And one predicate:

* IPAddr::Base.valid_prefix_length?